+++ build-release.ps1
@@ -1,5 +1,5 @@
 param(
-    [string]$Version = "3.0.3",
+    [string]$Version = "3.0.5",
     [string]$OutputDirectory
 )